slick integration tests
