+++ .gitignore
@@ -8,6 +8,6 @@ target
 
 # we don't want to push this file,
 # each contributor must create its own to signal acceptance
-akka-projection-jdbc/src/test/resources/container-license-acceptance.txt
-akka-projection-slick/src/test/resources/container-license-acceptance.txt
+akka-projection-jdbc/src/it/resources/container-license-acceptance.txt
+akka-projection-slick/src/it/resources/container-license-acceptance.txt
