--- akka-projection-jdbc/src/it/resources/container-license-acceptance.txt
@@ -1,1 +1,0 @@
-mcr.microsoft.com/mssql/server:2017-CU12
